Apply CR suggestions
